doc/noc-troubleshoot-assistant/mcp-server/requirements.txt
flask==3.0.0
gunicorn==21.2.0
zabbix-utils==2.0.0
pyyaml==6.0.1
requests>=2.31.0
